feat: Major components package restructuring. Migrated a lot of components into stateless. changelog: - added new packages under component package(alert, cards, lazy, rows, and textfield). - populating Constant.kt with long strings, now Kotlin file is less string-y.
