+++ app/src/main/java/id/arvigo/arvigobasecore/ui/component/alert/AlertLogout.kt
@@ -1,0 +1,41 @@
+package id.arvigo.arvigobasecore.ui.component.alert
+
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.MutableState
+
+@Composable
+fun AlertLogout(
+    openDialog : MutableState<Boolean>,
+    onClick: () -> Unit
+) {
+    AlertDialog(
+        onDismissRequest = {
+            openDialog.value = false
+        },
+        title = {
+            Text(text = "Logout")
+        },
+        text = {
+            Text(text = "Apakah kamu yakin bahwa kamu ingin logout?")
+        },
+        confirmButton = {
+            TextButton(
+                onClick = onClick
+            ) {
+                Text("Logout")
+            }
+        },
+        dismissButton = {
+            TextButton(
+                onClick = {
+                    openDialog.value = false
+                }
+            ) {
+                Text("Tidak")
+            }
+        }
+    )
+}
