+++ app/src/main/java/id/arvigo/arvigobasecore/ui/component/alert/AlertStateless.kt
@@ -1,0 +1,57 @@
+package id.arvigo.arvigobasecore.ui.component.alert
+
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Info
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.MutableState
+import androidx.compose.ui.text.style.TextAlign
+
+@Composable
+fun AlertStateless(
+    openDialog: MutableState<Boolean>,
+    confirmButton: @Composable () -> Unit,
+    title: String,
+    desc: String
+) {
+    AlertDialog(
+        onDismissRequest = {
+            openDialog.value = false
+        },
+        icon = {
+            Icon(
+                imageVector = Icons.Filled.Info,
+                contentDescription = "Info",
+                tint = MaterialTheme.colorScheme.primary
+            )
+        },
+        title = {
+            Text(
+                text = title,
+                style = MaterialTheme.typography.headlineSmall,
+                textAlign = TextAlign.Center,
+            )
+        },
+        text = {
+            Text(
+                text = desc,
+                style = MaterialTheme.typography.bodyLarge,
+                textAlign = TextAlign.Center,
+            )
+        },
+        confirmButton = confirmButton,
+        dismissButton = {
+            TextButton(
+                onClick = {
+                    openDialog.value = false
+                }
+            ) {
+                Text("Kembali")
+            }
+        }
+    )
+}
