+++ app/src/main/java/id/arvigo/arvigobasecore/ui/feature/wishlist/screen/ProductWishList.kt
@@ -3,7 +3,7 @@ package id.arvigo.arvigobasecore.ui.feature.wishlist.screen
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
-import id.arvigo.arvigobasecore.ui.component.ProductLazyGrid
+import id.arvigo.arvigobasecore.ui.component.lazy.ProductLazyGrid
 import id.arvigo.arvigobasecore.ui.feature.wishlist.ProductsUiState
 import id.arvigo.arvigobasecore.ui.feature.wishlist.model.WishListViewModel
 import id.arvigo.arvigobasecore.ui.theme.ArvigoBaseCoreTheme
